Adding 2 endpoints to get the schema
