+++ lib/controllers/app.js
@@ -16,6 +16,8 @@ module.exports = (req, res) => {
     resource.link('crud', RoutesInfo.expand('w2-app:crud'));
     resource.link('home', RoutesInfo.expand('w2-app:app', resource));
     resource.link('domain', RoutesInfo.expand('w2-app:domain', resource));
+    resource.link('schemaDomain', RoutesInfo.expand('w2-app:schema-domain', resource));
+    resource.link('schemaType', RoutesInfo.expand('w2-app:schema-type', resource));
 
     res.format({
         html: () => {
