+++ lib/controllers/index.js
@@ -3,11 +3,13 @@ const crud = require('./crud');
 const domain = require('./domain');
 const domainList = require('./domain-list');
 const portal = require('./portal');
+const schema = require('./schema');
 
 module.exports = {
     app,
     crud,
     domain,
     domainList,
-    portal
+    portal,
+    schema
 };
