+++ lib/routes.js
@@ -13,6 +13,12 @@ module.exports = (baseUrl) => {
     routesInfo.route('w2-app:domain', '/app/{domain}')
         .get(controllers.domain);
 
+    routesInfo.route('w2-app:schema-domain', '/app/schema/{domain}')
+        .get(controllers.schema.domain);
+
+    routesInfo.route('w2-app:schema-type', '/app/schema/{domain}/{type}')
+        .get(controllers.schema.type);
+
     routesInfo.route('w2-app:app', '/app/{domain}/{type}{?oid}')
         .get(controllers.app);
